+++ backend/src/test/java/com/NBE4_5_SukChanHoSu/BE/domain/user/controller/UserProfileControllerTest.java
@@ -100,205 +100,4 @@ class UserProfileControllerTest {
 //                .andExpect(jsonPath("$.code").value("201"));
 //    }
 
-    private String jwtToken;
-
-    @BeforeEach
-    void setUp() {
-        objectMapper = new ObjectMapper();
-        login();
-    }
-
-    @DisplayName("로그인")
-    void login() {
-        // given
-        String email = "initUser1@example.com";
-        String rawPassword = "testPassword123!";
-
-        // 로그인
-        UserLoginRequest loginDto = new UserLoginRequest();
-        loginDto.setEmail(email);
-        loginDto.setPassword(rawPassword);
-
-        // when
-        LoginResponse tokenDto = userService.login(loginDto);
-        this.jwtToken = tokenDto.getAccessToken();
-    }
-
-    @Test
-    @DisplayName("범위 내에 존재하는 사용자 조회")
-    void getUserWithinRadius() throws Exception {
-        //given
-        UserProfile userProfile = matchingService.findUser(1L);
-        int radius = userProfile.getSearchRadius();
-
-        //when
-        ResultActions action = mvc.perform(get("/api/matching/withinRadius")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-
-        // 파싱
-        String responseBody = action.andReturn().getResponse().getContentAsString();
-        JSONObject jsonResponse = new JSONObject(responseBody);
-        JSONArray usersArray = jsonResponse.getJSONArray("data");
-
-        // then
-        action.andExpect(status().isOk())
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message",containsString("성공")));
-
-        for (int i = 0; i < usersArray.length(); i++) {
-            JSONObject user = usersArray.getJSONObject(i);
-            String distanceStr = user.getString("distance");
-
-            // 거리 값에서 숫자만 추출
-            int distanceValue = extractDistanceValue(distanceStr);
-
-            // 반경(radius) 이하인지 확인
-            assertTrue(distanceValue <= radius);
-
-        }
-
-    }
-
-    // 거리 문자열에서 숫자만 추출하여 반환 (예: "약 1km" -> 1)
-    private int extractDistanceValue(String distance) {
-        String distanceNumber = distance.replaceAll("[^0-9]", "");
-        return Integer.parseInt(distanceNumber);
-    }
-
-    @Test
-    @DisplayName("범위 조절")
-    void updateRadius() throws Exception {
-        // given
-        int radius = 10;
-
-        // when
-        ResultActions action = mvc.perform(put("/api/matching/radius")
-                        .param("radius", String.valueOf(radius))
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-
-        // then
-        action.andExpect(status().isOk())
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message",containsString("성공")))
-                .andExpect(jsonPath("$.data.searchRadius").value(radius));
-
-        mvc.perform(get("/api/profile/profile/me")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message",containsString("성공")))
-                .andExpect(jsonPath("$.data.searchRadius").value(radius));
-    }
-
-    @Test
-    @DisplayName("이성 조회(거리 포함)")
-    void findProfileByGender() throws Exception {
-        // given
-
-        // when
-        ResultActions action = mvc.perform(get("/api/matching/gender")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-
-        // then
-        action.andExpect(status().isOk())
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message",containsString("성공")))
-                .andExpect(jsonPath("$.data[*].distance").exists());
-
-    }
-
-    @Test
-    @DisplayName("태그로 프로필 조회")
-    void findProfileByTags() throws Exception {
-        // given
-        UserProfile user =matchingService.findUser(1L);
-        List<Genre> tags = user.getFavoriteGenres();
-
-        // when
-        ResultActions action = mvc.perform(get("/api/matching/tags")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-        // 파싱
-        String responseBody = action.andReturn().getResponse().getContentAsString();
-        JSONObject jsonResponse = new JSONObject(responseBody);
-        JSONArray users = jsonResponse.getJSONArray("data");
-
-        // then
-        action.andExpect(status().isOk())
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message",containsString("성공")));
-
-        // 태그 겹치는지 확인
-        for (int i = 0; i < users.length(); i++) {
-            JSONObject userProfile = users.getJSONObject(i);
-            JSONArray favoriteGenres = userProfile.getJSONArray("favoriteGenres");
-
-            // 장르 비교
-            boolean hasMatchingGenre = false;
-            for (int j = 0; j < favoriteGenres.length(); j++) {
-                Genre genre = Genre.valueOf(favoriteGenres.getString(j));
-                if (tags.contains(genre)) {
-                    hasMatchingGenre = true;
-                    break;
-                }
-            }
-
-            // 하나라도 겹치지 않는 경우 실패
-            if (!hasMatchingGenre) {
-                fail("겹치지 않는 유저가 발견: " + userProfile.toString());
-            }
-        }
-    }
-
-    @Test
-    @DisplayName("추천")
-    void recommend() throws Exception {
-        // given
-        UserProfile user =matchingService.findUser(1L);
-
-        // when
-        // 추천 1
-        ResultActions action1 = mvc.perform(get("/api/matching/recommend")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-        // 응답 파싱
-        int status1 = action1.andReturn().getResponse().getStatus();
-        String responseBody1 = action1.andReturn().getResponse().getContentAsString();
-        JSONObject jsonResponse1 = new JSONObject(responseBody1);
-
-        // 추천 2
-        ResultActions action2 = mvc.perform(get("/api/matching/recommend")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + jwtToken))
-                .andDo(print());
-        // 응답 파싱
-        int status2 = action2.andReturn().getResponse().getStatus();
-        String responseBody2 = action2.andReturn().getResponse().getContentAsString();
-        JSONObject jsonResponse2 = new JSONObject(responseBody2);
-
-        // 404 체크
-        if (status1 == HttpStatus.NOT_FOUND.value()) {
-            assertEquals("404", jsonResponse1.getString("code"));
-            assertEquals("추천할 사용자가 없습니다.", jsonResponse1.getString("message"));
-        } else if (status2 == HttpStatus.NOT_FOUND.value()) {
-            assertEquals("404", jsonResponse2.getString("code"));
-            assertEquals("추천할 사용자가 없습니다.", jsonResponse2.getString("message"));
-        } else{
-            // 둘다 200 OK를 반환한 경우 ->  응답이 달라야함
-            JSONObject user1 = jsonResponse1.getJSONObject("data");
-            JSONObject user2 = jsonResponse2.getJSONObject("data");
-
-            assertNotEquals(user1.toString(), user2.toString());
-        }
-
-    }
-
 }
